:truck: move utils about numeric to certif contexte
