+++ api/tests/unit/domain/read-models/EuropeanNumericLevelFactory_test.js
@@ -1,5 +1,5 @@
-import { EuropeanNumericLevel } from '../../../../src/shared/domain/read-models/EuropeanNumericLevel.js';
-import { EuropeanNumericLevelFactory } from '../../../../src/shared/domain/read-models/EuropeanNumericLevelFactory.js';
+import { EuropeanNumericLevel } from '../../../../src/certification/session-management/domain/read-models/EuropeanNumericLevel.js';
+import { EuropeanNumericLevelFactory } from '../../../../src/certification/session-management/domain/read-models/EuropeanNumericLevelFactory.js';
 import { expect } from '../../../test-helper.js';
 
 describe('Unit | Domain | Read-models | EuropeanNumericLevelFactory', function () {
